refactor: 기록 앱바, 바디 위젯 분리
